+++ rpcs3/Loader/iso_validation.h
