+++ rpcs3/emucore.vcxproj.filters
@@ -1429,7 +1429,7 @@
     <ClCompile Include="Emu\RSX\Overlays\overlay_select.cpp">
       <Filter>Emu\GPU\RSX\Overlays</Filter>
     </ClCompile>
-    <ClCompile Include="Loader\ISO_validation.cpp">
+    <ClCompile Include="Loader\iso_validation.cpp">
       <Filter>Loader</Filter>
     </ClCompile>
   </ItemGroup>
@@ -2878,7 +2878,7 @@
     <ClInclude Include="Emu\RSX\Common\aligned_malloc.hpp">
       <Filter>Emu\GPU\RSX\Common</Filter>
     </ClInclude>
-    <ClInclude Include="Loader\ISO_validation.h">
+    <ClInclude Include="Loader\iso_validation.h">
       <Filter>Loader</Filter>
     </ClInclude>
   </ItemGroup>
